LayoutTests/platform/wk2/http/tests/plugins/get-url-redirect-notify-expected.txt


Sync allow, reason shouldn't be NPRES_USER_BREAK...true
Sync allow, URL should be the final URL...true
Sync deny, reason should be NPRES_USER_BREAK...true
Sync deny, URL should be the previous URL...true
Async allow, reason shouldn't be NPRES_USER_BREAK...true
Async allow, URL should be the final URL...true
Async deny, reason should be NPRES_USER_BREAK...true
Async deny, URL should be the previous URL...true
Async multiple allow, reason shouldn't be NPRES_USER_BREAK...true
Async multiple allow, URL should be the final URL...true
Async multiple deny, reason should be NPRES_USER_BREAK...true
Async multiple deny, URL should be the previous URL...true
